+++ spark-bigquery-connector-common/src/main/java/com/google/cloud/spark/bigquery/SparkBigQueryConfig.java
@@ -180,7 +180,7 @@ public class SparkBigQueryConfig
   // used to create BigQuery ReadSessions
   private com.google.common.base.Optional<String> traceId;
   private ImmutableMap<String, String> bigQueryJobLabels = ImmutableMap.of();
-  private com.google.common.base.Optional<Long> readSessionTimeoutInSeconds;
+  private com.google.common.base.Optional<Long> createReadSessionTimeoutInSeconds;
 
   @VisibleForTesting
   SparkBigQueryConfig() {
@@ -439,15 +439,9 @@ public class SparkBigQueryConfig
 
     config.bigQueryJobLabels = parseBigQueryJobLabels(globalOptions, options);
 
-    config.readSessionTimeoutInSeconds =
-        getAnyOption(globalOptions, options, "readSessionTimeoutInSeconds")
+    config.createReadSessionTimeoutInSeconds =
+        getAnyOption(globalOptions, options, "createReadSessionTimeoutInSeconds")
             .transform(Long::parseLong);
-    if (config.readSessionTimeoutInSeconds.isPresent()
-        && config.readSessionTimeoutInSeconds.get() < 0) {
-      throw new IllegalArgumentException(
-          "readSessionTimeoutInSeconds must have a positive value, the configured value is "
-              + config.readSessionTimeoutInSeconds);
-    }
     return config;
   }
 
@@ -735,8 +729,8 @@ public class SparkBigQueryConfig
   }
 
   @Override
-  public Optional<Long> getReadSessionTimeoutInSeconds() {
-    return readSessionTimeoutInSeconds.toJavaUtil();
+  public Optional<Long> getCreateReadSessionTimeoutInSeconds() {
+    return createReadSessionTimeoutInSeconds.toJavaUtil();
   }
 
   @Override